Update
Created Skills Class
updated Student and Project to use Skills
application now creates random set of Students and Projects.

Will work most of the time. sometimes if we can't find students for a
project, such as a local student for a local project, we will get thrown
an instance of std::out_of_range. May need to make team size dynamic. If
we can't find a student to fill a project team slot, we need to do
something.