Update
fixed issue with out of bounds exception when a team could not be filled
for a project. fixed by checking for out of bounds, in this case an
initialized value of -1. this would indicate that project team of 5
couldn't be reached for a project. action was to move on to next
project.

Testing will involve using different number of students and projects.
This is currently tested with 20 students and 4 projects. Both students
and projects are randomly generated.