In US45Task65/src/US45Task65.cpp
Added functionality to compute all the combinations of students in teams and using their combined skill score for a project.

The calculation of combinations is done iteratively, not recursively. Also no variable is used to store which iteration it's on.

So you don't have to worry about running out of memory when running through all the combinations. The only constraint is time.